Update README.md
As environmental, social, and governance (ESG) becomes an increasingly important initiative for customers, stakeholders, and regulatory authorities worldwide, finance teams shift their focus to collecting and reporting on ESG data. To accomplish this, they need to identify which information is essential to meeting standards and then develop a system for collecting it from multiple sources. However, the fragmented state of ESG data – and the fact that most of this data is collected manually – complicates the task, demanding additional time and resources to find a resolution.
Reconciling complex data, ensuring auditability, and keeping up with evolving standards is a tremendous undertaking. This underscores the need for finance teams to develop a strategy for incorporating ESG into their financial planning and reporting activities and find solutions that help them on their journey.
READ MORE-https://www.iconresources.com/navigating-esg-data-management-in-a-complex-world/